Feature: Support G5AR Gateway (#13)
T-Mobile's G5AR Gateway supports 5G Standalone (SA) mode, and as a result, 4G signal parameters are not always available.

In this version, all 4G and 5G signal entities are disabled by default, for performance as well as for the above reason. Users can enable them as needed.

Also, controlling 6G Wi-Fi is now supported if the gateway supports it.

Other adjustments were made for features not supported by the G5AR.